Support manually editing bone bind strategies.
In order to be able to fit a rig to a modified character mesh,
when saving it to a json file MPFB has to bind it by choosing
one of the three strategies for each bone head and tail, and
the associated parameters.

In some cases this automatic binding may be sub-optimal.
Although the best way would be to have a proper UI for manual
binding adjustment, it is possible to simply edit the json
file. However, if the rig is loaded and re-saved, it will
overwrite all manual changes with automatic ones.

This commit saves the bindings from the file in custom properties
of the bones when loading a rig for development, and reuses them
when saving if the automatic choice doesn't give a better match.
The intent is that bindings shouldn't change if bones weren't moved.