Use the new style face in the Rigify metarig and tweak option defaults.
Since actually including the face in the metarig is rather
complicated, for now use the conversion operator to upgrade
the currently used legacy face.

The new face includes deform bones for teeth, so switch the
default for the teeth option to do nothing. Also, deleting the
metarig goes against the workflow of Rigify, which includes
the ability to re-generate in place after tweaking metarig.