Apply Preserve Volume to fingers in the Rigify rig.
Preserve Volume cannot be applied globally because it has
serious unnatural effects in certain areas (e.g. shoulders),
and custom helpers allow applying extra transformations.
However for fingers it is the most appropiate choice.

This requires code changes to support masks when saving and
loading weights, and to add a second Armature modifier.