Allow mhclo assets to have their own child sub-rig layers.
In order to handle cases where clothing has features that
are reasonable to make animateable, but cannot be covered by
the main skeleton (e.g. a ponytail in a hair asset), it is
necessary to allow assets to include rigs of their own.

When a sub-rig is loaded, it is inserted in the parenting
hierarchy between the main skeleton and the asset mesh, but
attaching its bones to the main skeleton has to be done via
constraints (e.g. Child Of) in the sub-rig itself that reference
bones from the main skeleton. Only bones that bridge two joint
cubes can be targeted in this way.

Sub-rig deformation is layered on top of the main rig via
the Multi-Modifier option of the Armature modifier and the
'mhmask-subrig' vertex group that specifies areas controlled
by the sub-rig. Disabling the modifier restores full control
to the main skeleton rig.

The sub-rig is stored in "NAME.rig.json" next to "NAME.mhclo"
using the regular format used for the main skeletons. The only
difference is that sub-rigs use joint locations from the base
mesh (while vertices are taken from the asset mesh), and their
constraints can use main skeleton bones as targets. In order
to avoid tying the asset rig to a specific skeleton, these
bones are identified by the two joint cubes they bridge.

Sub-rigs can be Rigify-ready (i.e. have appropriate settings
for meaningful generation), in which case the Generate button
from the Rigging panel will work for it. If after generation
the metarig is retained, the Refit Assets button will attempt
to refit the metarig and automatically re-generate. It is
recommended to provide an appropriately located 'root' bone
in a Rigify-ready sub-rig; it can be a good place for Child Of.