Introduce a MPFBSKEL file format that groups bone and rigify layer data.
This format is inspired by MHSKEL files, but is incompatible with them.
Rigify rigs are converted to this format.

Also delete operators for saving/loading rigify layers, add an option
to save rigify data with the rig, and an option to refit the rig after
saving to match the chosen strategies.