Implement UI for manual editing of bone fitting strategies.
In order to fit rigs to varying body meshes, MPFB binds bone end
points to vertices via a number of available strategies. Originally
the assignment was automatic, but when working on polishing the
Rigify rig, it was necessary to support manual assignment.

Previously this was achieved via manual editing of custom properties
of the rig bones and/or editing the saved json file plus reloading.
This is obviously highly technical and inconvenient.

This commit adds a UI panel for doing that. The panel appears in
the Bone tab of Properties in the edit mode of 'developer mode' rigs.
In order to mark a rig as developer mode, it is necessary to either
load it through the Developer panel, or save it through the same
panel, using the 'refit' option. In addition to enabling the mode,
this stores the current values of the strategies into bone properties.

The panel shows the current values of the strategies, and allows
changing between strategies. For strategies that use vertex lists,
it provides buttons to switch between editing the armature and
corresponding mesh (the properties are pinned while in the mesh mode),
and transferring data between mesh vertex selection and the strategy.

To facilitate editing consistent connected bone chains, an option
to keep connected bone ends in sync is provided and enabled by default.
Depending on the operation, connected ends are detected by comparing
strategies and/or exact locations. Additionally, a button will copy
strategies from connected bones to the current one.

One major caveat is that due to a long standing bug in the Blender
undo system, any edits to strategy properties done in mesh edit mode
cannot be reverted except by undoing all the way to reach armature edit
before the switch to mesh edit. Because of this, the operators provided
for dealing with the mesh only record undo points at the point of switch
in order to minimize the number of unproductive undo steps in mesh mode.