Record extra generated Rigify deform bones in the rig data.
Rigify can generate more deform bones than present in the metarig,
which causes a problem for filtering 'relevant' vertex groups when
loading weights.

Previously this was handled by allowing all DEF groups. This commit
adds an explicit list that is stored in the rig data file, and loaded
into a property of the armature object. This property is then used
by weight loading code.

Rigify rigs are updated to include this data, and all=True is removed.