Support an arbitrary offset in the bone end fitting strategy.
Some bones in the Rigify face are ideally supposed to be aligned
to an axis. In order to support that, allow including an arbitrary
offset vector in the fitting strategies. In this way the bone can
be aligned by making the tail use the same strategy as the head,
but with an offset along one axis.

The new offset setting is appropriately included in the UI panel.