Convert serialized Rigify rigs to the new face format.
Basically, the result of applying the convert to new face operator,
and assigning the appropriate strategies to the newly added bones
to match the surroundings.

The nose master and face bones use the strategy offset feature
to achieve perfect alignment to a world axis.

With this change, Refit Assets to Basemesh can work even after
generating the rigify rig (provided the metarig isn't removed)
by refitting the metarig and automatically re-generating.
Previously it failed because it didn't know how to refit the
metarig bones created by converting to the new face.